# gest-ashoex-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Health Check",
        "description": "API para verificar el estado de salud del sistema",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/edificios": {
            "get": {
                "tags": [
                    "Edificios"
                ],
                "summary": "Obtener todos los edificios con sus ubicaciones",
                "description": "Devuelve una lista de todos los edificios, incluyendo sus ubicaciones asociadas",
                "operationId": "obtenerEdificios",
                "responses": {
                    "200": {
                        "description": "Lista de edificios obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id_edificio": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nombre_edificio": {
                                                "type": "string",
                                                "example": "Edificio A"
                                            },
                                            "geolocalizacion": {
                                                "type": "string",
                                                "example": "Latitud: 19.4326, Longitud: -99.1332"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-10-30T12:34:56Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-10-30T12:34:56Z"
                                            },
                                            "ubicaciones": {
                                                "description": "Lista de ubicaciones dentro del edificio",
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id_ubicacion": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "piso": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "id_edificio": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": "2024-10-30T12:34:56Z"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": "2024-10-30T12:34:56Z"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error interno del servidor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Edificios"
                ],
                "summary": "Crear un nuevo edificio",
                "description": "Crea un edificio con el nombre y la geolocalización proporcionados.",
                "operationId": "createEdificio",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre_edificio"
                                ],
                                "properties": {
                                    "nombre_edificio": {
                                        "type": "string",
                                        "example": "Edificio Central"
                                    },
                                    "geolocalizacion": {
                                        "type": "string",
                                        "example": "-19.57347,-65.75537"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Éxito. Edificio creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nombre_edificio": {
                                            "type": "string",
                                            "example": "Edificio Central"
                                        },
                                        "geolocalizacion": {
                                            "type": "string",
                                            "example": "-19.57347,-65.75537"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-10-28T12:34:56Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-10-28T12:34:56Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación. Faltan datos o no son válidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "El nombre del edificio es obligatorio."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}